pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
local_scheme = "no-local-version"

[project]
name = "dbus2mqtt"
# version = "0.1.0"
dynamic = ["version"]
description = """\
    A Python tool to expose Linux D-Bus signals, methods and properties over MQTT \
    - featuring templating, payload enrichment and Home Assistant-ready examples \
    """
readme = "README.md"
license = "MIT"
keywords = ["python", "mqtt", "mpris", "dbus", "home-assistant"]
classifiers = [
    # https://pypi.org/classifiers/
    "Development Status :: 4 - Beta",

    "Intended Audience :: Developers",
    "Operating System :: POSIX :: Linux",
    "License :: OSI Approved :: MIT License",

    # Supported Python versions
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

requires-python = ">=3.10"

dependencies = [
    "apscheduler>=3.11.0",
    "colorlog>=6.9.0",
    "dbus-fast>=2.44.1",
    "janus>=2.0.0",
    "jinja2>=3.1.6",
    "jinja2-ansible-filters>=1.3.2",
    "jsonargparse>=4.38.0",
    "paho-mqtt>=2.1.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "schedule>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "types-pyyaml>=6.0.12.20241230",
    "taskipy>=1.14.1",
    "ruff>=0.11.2",
    "pyright>=1.1.396",
    "pre-commit>=4.2.0",
    "pip>=25.0.1",
    "pytest-asyncio>=1.0.0"
]

[project.urls]
# Homepage = "https://example.com"
# Documentation = "https://readthedocs.org"
Repository = "https://github.com/jwnmulder/dbus2mqtt.git"
Issues = "https://github.com/jwnmulder/dbus2mqtt/issues"
# Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

[project.scripts]
dbus2mqtt = "dbus2mqtt.main:main"

[tool.taskipy.tasks]
lint-ruff = "ruff check"
lint-pyright = "pyright"

[tool.pytest.ini_options]
addopts = "-s"
testpaths = ["tests"]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"
asyncio_default_test_loop_scope = "function"
log_cli = true

[tool.pyright]
venvPath = "."
venv = ".venv"

typeCheckingMode = "basic"

[tool.ruff.lint]
extend-select = [

    # pycodestyle Error, Warnings
    "E",
    "W",

    # isort
    "I",

    # pep8-naming
    "N",

    # pyupgrade
    "UP"
]
ignore = [
    # line-too-long
    "E501"
]

[tool.ruff.lint.isort]
lines-between-types = 1
